servlet2

+++ beautysalon/src/main/java/com/test/model/dao/GenericDao.java
@@ -5,9 +5,9 @@ import java.util.Optional;
 
 public interface GenericDao<T> extends AutoCloseable {
     void create (T entity);
-    Optional<T> findById(int id);
+    Optional<T> findById(Long id);
     List<T> findAll();
     void update(T entity);
-    void delete(int id);
+    void delete(Long id);
     void close();
 }
